feat: aprimorei o cadastro de produto pra receber imagem

+++ src/controladores/produtos/excluirProduto.js
@@ -1,10 +1,24 @@
 const { StatusCodes } = require("http-status-codes");
-const { excluirPorID } = require("../../provedor/produtosQuerys/queryFuncoes");
-const { ErroNaoEncontrado } = require("../../uteis/erros/erroDaApi");
+const {
+  ErroNaoEncontrado,
+  ErroDeConflito,
+} = require("../../uteis/erros/erroDaApi");
 const { erroProdutoNaoEncontrado } = require("../../uteis/erros/mensagens");
+const {
+  excluirPorID,
+  procurarProdutosEmPedidos,
+} = require("../../provedor/produtosQuerys/queryFuncoes");
 
 const excluirProduto = async (req, res) => {
   const { id } = req.params;
+
+  const produtoNaoPodeExcluir = await procurarProdutosEmPedidos(id);
+
+  if (produtoNaoPodeExcluir) {
+    throw ErroDeConflito(
+      "O Produto está vinculado á algum pedido, não pode ser excluido."
+    );
+  }
 
   const produtoNaoExiste = await excluirPorID(id);
